+++ Sources/VCore/Services and Managers/Presentation Host/Presentation Implementation/PresentationHostHostingViewControllerContainerViewController.swift
@@ -10,7 +10,7 @@
 import SwiftUI
 
 // MARK: - Presentation Host Hosting View Controller Container View Controller
-final class PresentationHostHostingViewControllerContainerViewController: KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubviewHeight {
+final class PresentationHostHostingViewControllerContainerViewController: KeyboardResponsiveUIViewController {
     // MARK: Subviews
     let hostingController: HostingViewControllerType
 
@@ -22,10 +22,17 @@ final class PresentationHostHostingViewControllerContainerViewController: Keyboa
     // MARK: Initializers
     init(
         uiModel: PresentationHostUIModel,
-        hostingController: HostingViewControllerType
+        content: AnyView
     ) {
         self.uiModel = uiModel
-        self.hostingController = hostingController
+
+        self.hostingController = {
+            let hostingController: UIHostingController = .init(rootView: content)
+            hostingController.view.translatesAutoresizingMaskIntoConstraints = false
+            hostingController.view.backgroundColor = nil
+            hostingController._disableSafeAreas()
+            return hostingController
+        }()
 
         super.init(nibName: nil, bundle: nil)
     }
@@ -37,13 +44,26 @@ final class PresentationHostHostingViewControllerContainerViewController: Keyboa
     // MARK: Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
+        setUp()
+    }
 
+    // MARK: Setup
+    private func setUp() {
+        setUpView()
+        addSubviews()
+        setUpLayout()
+    }
+
+    private func setUpView() {
+        view.backgroundColor = nil
+    }
+
+    private func addSubviews() {
         addChild(hostingController)
-
-        hostingController.view.translatesAutoresizingMaskIntoConstraints = false
-
         view.addSubview(hostingController.view)
+    }
 
+    private func setUpLayout() {
         NSLayoutConstraint.activate([
             hostingController.view.constraintLeading(to: view),
             hostingController.view.constraintTrailing(to: view),
@@ -53,18 +73,123 @@ final class PresentationHostHostingViewControllerContainerViewController: Keyboa
     }
 
     // MARK: Keyboard Responsiveness
-    override var keyboardResponsivenessFirstResponderViewMarginBottomToKeyboard: CGFloat { uiModel.focusedViewKeyboardSafeAreaInset }
-
     override func keyboardWillShow(_ systemKeyboardInfo: SystemKeyboardInfo) {
-        guard uiModel.handlesKeyboardResponsiveness else { return }
-
         super.keyboardWillShow(systemKeyboardInfo)
+
+        switch uiModel.keyboardResponsivenessStrategy?._keyboardResponsivenessStrategy {
+        case nil:
+            break
+
+        case .offset(let offset):
+            UIView.animateKeyboardResponsiveness(
+                systemKeyboardInfo: systemKeyboardInfo,
+                animations: { [weak self] in
+                    guard let self else { return }
+
+                    view.bounds.origin.y = offset
+                    view.superview?.layoutIfNeeded()
+                }
+            )
+
+        case .offsetByKeyboardHeight(let additionalOffset):
+            UIView.animateKeyboardResponsiveness(
+                systemKeyboardInfo: systemKeyboardInfo,
+                animations: { [weak self] in
+                    guard let self else { return }
+
+                    guard let systemKeyboardHeight: CGFloat = systemKeyboardInfo.frame?.size.height else { return } // Will never fail
+
+                    view.bounds.origin.y = systemKeyboardHeight + additionalOffset
+                    view.superview?.layoutIfNeeded()
+                }
+            )
+
+        case .offsetByObscuredSubviewHeight(let safeAreaInset):
+            guard let firstResponderSubview: UIView = view.childFirstResponderView else { return }
+
+            UIView.animateKeyboardResponsivenessByUnObscuringFirstResponderView(
+                keyboardWillShow: true,
+                firstResponderView: firstResponderSubview,
+                containerView: view,
+                systemKeyboardInfo: systemKeyboardInfo,
+                keyboardSafeAreMargin: safeAreaInset
+            )
+        }
     }
 
     override func keyboardWillHide(_ systemKeyboardInfo: SystemKeyboardInfo) {
-        guard uiModel.handlesKeyboardResponsiveness else { return }
-
         super.keyboardWillHide(systemKeyboardInfo)
+
+        switch uiModel.keyboardResponsivenessStrategy?._keyboardResponsivenessStrategy {
+        case nil:
+            break
+
+        case .offset:
+            UIView.animateKeyboardResponsiveness(
+                systemKeyboardInfo: systemKeyboardInfo,
+                animations: { [weak self] in
+                    guard let self else { return }
+
+                    view.bounds.origin.y = 0
+                    view.superview?.layoutIfNeeded()
+                }
+            )
+
+        case .offsetByKeyboardHeight:
+            UIView.animateKeyboardResponsiveness(
+                systemKeyboardInfo: systemKeyboardInfo,
+                animations: { [weak self] in
+                    guard let self else { return }
+
+                    view.bounds.origin.y = 0
+                    view.superview?.layoutIfNeeded()
+                }
+            )
+
+        case .offsetByObscuredSubviewHeight(let safeAreaInset):
+            guard let firstResponderSubview: UIView = view.childFirstResponderView else { return }
+
+            UIView.animateKeyboardResponsivenessByUnObscuringFirstResponderView(
+                keyboardWillShow: false,
+                firstResponderView: firstResponderSubview,
+                containerView: view,
+                systemKeyboardInfo: systemKeyboardInfo,
+                keyboardSafeAreMargin: safeAreaInset
+            )
+        }
+    }
+}
+
+// MARK: - Helpers
+extension UIHostingController {
+    // https://github.com/scenee/FloatingPanel/issues/454
+    // https://defagos.github.io/swiftui_collection_part3
+    fileprivate func _disableSafeAreas() {
+        guard let viewClass: AnyClass = object_getClass(view) else { return }
+
+        let viewSubclassName: String = String(cString: class_getName(viewClass)).appending("_IgnoreSafeArea")
+
+        if let viewSubclass: AnyClass = NSClassFromString(viewSubclassName) {
+            object_setClass(view, viewSubclass)
+
+        } else if
+            let viewClassNameUtf8: UnsafePointer<CChar> = (viewSubclassName as NSString).utf8String,
+            let viewSubclass: AnyClass = objc_allocateClassPair(viewClass, viewClassNameUtf8, 0)
+        {
+            if let method: Method = class_getInstanceMethod(UIView.self, #selector(getter: UIView.safeAreaInsets)) {
+                let safeAreaInsets: @convention(block) (AnyObject) -> UIEdgeInsets = { _ in .zero }
+
+                class_addMethod(
+                    viewSubclass,
+                    #selector(getter: UIView.safeAreaInsets),
+                    imp_implementationWithBlock(safeAreaInsets),
+                    method_getTypeEncoding(method)
+                )
+            }
+
+            objc_registerClassPair(viewSubclass)
+            object_setClass(view, viewSubclass)
+        }
     }
 }