+++ Sources/VCore/Views/UIKit/Keyboard Responsive UI View Controller/FirstResponderViewUnObscuringUIViewController.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubviewHeight.swift
+//  FirstResponderViewUnObscuringUIViewController.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 15.07.22.
@@ -9,10 +9,10 @@
 
 import UIKit
 
-// MARK: - Keyboard Responsive UI View Controller Offsetting Container by Obscured Subview Height
-/// Subclass of  `KeyboardResponsiveUIViewController` that handles keyboard notifications by offsetting container by the obscured subview height.
+// MARK: - First Responder View Un-Obscuring UI View Controller
+/// Subclass of  `KeyboardResponsiveUIViewController` that handles keyboard notifications by  un-obscuring first responder view, if needed.
 @available(tvOS, unavailable)
-open class KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubviewHeight: KeyboardResponsiveUIViewController {
+open class FirstResponderViewUnObscuringUIViewController: KeyboardResponsiveUIViewController {
     // MARK: Properties
     /// `UIView` on which offset animations will be applied. Set to `view`.
     ///
@@ -21,8 +21,8 @@ open class KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubvie
     /// In case of `UIScrollView`, content view can be used.
     open var keyboardResponsivenessContainerView: UIView { view }
 
-    /// Bottom margin from first responder view to keyboard. Set to `20`.
-    open var keyboardResponsivenessFirstResponderViewMarginBottomToKeyboard: CGFloat { 20 }
+    /// Keyboard safe area margin from first responder view to keyboard. Set to `20`.
+    open var keyboardResponsivenessFirstResponderViewKeyboardSafeAreaMargin: CGFloat { 20 }
     
     // MARK: Keyboard Responsiveness
     open override func keyboardWillShow(_ systemKeyboardInfo: SystemKeyboardInfo) {
@@ -30,12 +30,12 @@ open class KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubvie
         
         guard let firstResponderSubview: UIView = view.childFirstResponderView else { return }
         
-        UIView.animateKeyboardResponsivenessByOffsettingContainerByObscuredSubviewHeight(
+        UIView.animateKeyboardResponsivenessByUnObscuringFirstResponderView(
             keyboardWillShow: true,
             firstResponderView: firstResponderSubview,
             containerView: keyboardResponsivenessContainerView,
             systemKeyboardInfo: systemKeyboardInfo,
-            marginBottom: keyboardResponsivenessFirstResponderViewMarginBottomToKeyboard
+            keyboardSafeAreMargin: keyboardResponsivenessFirstResponderViewKeyboardSafeAreaMargin
         )
     }
     
@@ -44,12 +44,12 @@ open class KeyboardResponsiveUIViewControllerOffsettingContainerByObscuredSubvie
         
         guard let firstResponderSubview: UIView = view.childFirstResponderView else { return }
         
-        UIView.animateKeyboardResponsivenessByOffsettingContainerByObscuredSubviewHeight(
+        UIView.animateKeyboardResponsivenessByUnObscuringFirstResponderView(
             keyboardWillShow: false,
             firstResponderView: firstResponderSubview,
             containerView: keyboardResponsivenessContainerView,
             systemKeyboardInfo: systemKeyboardInfo,
-            marginBottom: keyboardResponsivenessFirstResponderViewMarginBottomToKeyboard
+            keyboardSafeAreMargin: keyboardResponsivenessFirstResponderViewKeyboardSafeAreaMargin
         )
     }
 }